prevent invalidating an outdated renderable
in case someone leaks the controller and calls invalidate on it later,
only allow this if we're still rendering the associated renderable

thought: should this error, or just silently ignore the operation?
I think it should throw, because you can always ignore it, and it likely signifies
a programmer error.