add a test to check compatibility with custom elements
remembered a Rich tweet that described a bug svelte had in their handling
of custom elements, and wrote a test to make sure we don't have the same issue.

as it turns out, because we append the fragment to the dom before we run the patches,
we actually avoid this problem, but now we have a test to make sure this behavior
never regresses.
https://x.com/Rich_Harris/status/1841467510194843982