switch to a non-live reimplementation of Ranges
this should drastically help with performance, because now we
don't have a ton of live ranges on the page that have to be
modified for every single dom change, and also it fixes the
bug in 62a78908 because the live tracking got in our way.

instead, we do the accounting that the dom would be doing
ourselves. (or at least the parts we actually want)