fix a bug related to wrong ChildPart spans (#40)
+ add failing test

in html`${a}|${b}`,
a's slot's childIndex is 0
b's slot's childIndex is 2

however, where ${a} results in 2+ nodes:
by the time ${b} is created, it's no longer at the same
index in the parent.

by reversing the order in which the parts are created,
the slots are always in the expected position.