fix invalidating after hiding and showing a renderable
i can't reduce this to write a test, but the "loading..." flicker of
the kanban example before https://github.com/tombl/dhtml/pull/70 would
do something weird with the renderable identity, and I would
simultaneously get a "renderable has not been rendered" error and a
"this is not the current renderable" error.

this fixes it by always updating the _invalidate function on the
controller, so the method doesn't hold onto a stale renderable.