fix span
you can have more than one css directive on an element,
so just the dynamic index is not enough to make a unique name.

future question: do two instantiations of the same template
on the same element need different names?
probably not, because they will have duplicate styles so
the overlap is undefined behavior.